Refactoring: move few helpers to .util package
