+++ src/main/scala/tech/sourced/gemini/FeaturesHash.scala
@@ -1,6 +1,7 @@
 package tech.sourced.gemini
 
 import tech.sourced.featurext.generated.service.Feature
+import tech.sourced.gemini.util.MathUtil
 
 import scala.collection.Searching._
