+++ src/main/scala/tech/sourced/gemini/OrderedDocFreq.scala
@@ -8,6 +8,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
 import com.fasterxml.jackson.module.scala.experimental.ScalaObjectMapper
 
+
 /**
   * Ordered document frequency
   *
