+++ src/main/scala/tech/sourced/gemini/cmd/QueryApp.scala
@@ -4,7 +4,8 @@ import com.datastax.driver.core.Cluster
 import io.grpc.ManagedChannelBuilder
 import org.bblfsh.client.BblfshClient
 import tech.sourced.featurext.generated.service.FeatureExtractorGrpc
-import tech.sourced.gemini.{Gemini, Logger, QueryResult}
+import tech.sourced.gemini.util.Logger
+import tech.sourced.gemini.{Gemini, QueryResult}
 
 case class QueryAppConfig(file: String = "",
                           host: String = Gemini.defaultCassandraHost,
