+++ src/test/scala/tech/sourced/gemini/FileQuerySpec.scala
@@ -13,6 +13,7 @@ import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
 import tags.DB
 import tech.sourced.featurext.generated.service.FeatureExtractorGrpc.FeatureExtractor
 import tech.sourced.featurext.generated.service._
+import tech.sourced.gemini.util.Logger
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.parsing.json.JSON
